error handling

+++ src/main/java/com/example/back/services/Impl/ClasseServiceImpl.java
@@ -8,7 +8,6 @@ import com.example.back.services.ClasseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +18,18 @@ public class ClasseServiceImpl implements ClasseService {
     private ClasseRepository classeRepository;
 
     @Override
-    public boolean create(Classe classe) {
+    public boolean create(Classe classe) throws Exception {
+        return saveCheck(classe);
+    }
+
+    private boolean saveCheck(Classe classe) throws Exception {
+        if(classe.getNom() == null || classe.getNom().isEmpty()){
+            throw new Exception("important field empty");
+        }
+        Classe found = classeRepository.findByNom(classe.getNom());
+        if (found != null) {
+            throw new Exception("classe name already exists");
+        }
         classeRepository.save(classe);
         return true;
     }
@@ -30,19 +40,24 @@ public class ClasseServiceImpl implements ClasseService {
     }
 
     @Override
-    public ClasseDTO read(Integer id) {
+    public ClasseDTO read(Integer id) throws Exception {
         Optional<Classe> t = classeRepository.findById(id);
-        return new ClasseDTO(t.orElse(null));
+        return new ClasseDTO(t.orElseThrow(() -> new Exception("Classe id does not exist")));
     }
 
     @Override
-    public boolean update(Classe classe) {
-        classeRepository.save(classe);
-        return true;
+    public boolean update(Classe classe) throws Exception {
+        if(!classeRepository.existsById(classe.getId())){
+            throw new Exception("Classe id does not exist");
+        }
+        return saveCheck(classe);
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Integer id) throws Exception {
+        if(!classeRepository.existsById(id)){
+            throw new Exception("Classe id does not exist");
+        }
         classeRepository.deleteById(id);
         return true;
     }
@@ -50,5 +65,10 @@ public class ClasseServiceImpl implements ClasseService {
     @Override
     public List<ClasseDTO> searchByNom(String nom) {
         return DTOListMapper.mapClasse(classeRepository.findByNomContains(nom));
+    }
+
+    @Override
+    public ClasseDTO findByNom(String nom) {
+        return new ClasseDTO(classeRepository.findByNom(nom));
     }
 }
